refactor: 바텀 시트 관련 클래스명 수정 및 폴더링
