+++ lib/presentation/screen/component/color_selection_field.dart
@@ -1,8 +1,8 @@
-import 'package:calendar_scheduler/presentation/screen/component/schedule_register_bottom_sheet.dart';
 import 'package:flutter/material.dart';
 
 import '../../const/strings.dart';
 import '../../const/styles.dart';
+import '../schedule_regist/create_schedule_bottom_sheet.dart';
 
 class ColorSelectionField extends StatelessWidget {
   final List<Color> colors;
