+++ lib/presentation/screen/component/time_input_field.dart
@@ -1,5 +1,4 @@
 import 'package:calendar_scheduler/presentation/extension.dart';
-import 'package:calendar_scheduler/presentation/screen/component/schedule_register_bottom_sheet.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
@@ -7,6 +6,7 @@ import 'package:flutter/widgets.dart';
 import '../../const/colors.dart';
 import '../../const/strings.dart';
 import '../../const/styles.dart';
+import '../schedule_regist/create_schedule_bottom_sheet.dart';
 
 class TimeInputField extends StatefulWidget {
   final int initialTime;
